# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portfólio API Docs",
        "description": "API para consulta de dados do portfólio",
        "contact": {
            "name": "Felipe Silva",
            "email": "felipe.allware@gmail.com"
        },
        "license": {
            "name": "Licença API",
            "url": "https://github.com/felipesilva15/portfolio-manager-api/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log in",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "description": "Data for creating a new participant",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccessTokenDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logged in user data",
                "operationId": "9d9bc556f927575aa492a88c28fe5908",
                "responses": {
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logout",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "integer",
                                            "example": "Logout efetuado com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/refresh-token": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh the access token",
                "operationId": "41b109d150a67fd6df88cced4d491a6f",
                "responses": {
                    "200": {
                        "description": "Token details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccessTokenDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/contact": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "List all contacts",
                "operationId": "5a13914944335335387a39ef5445925e",
                "responses": {
                    "200": {
                        "description": "Contact list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contact"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Contact"
                ],
                "summary": "Registers a contact",
                "operationId": "4d33e9ce28a28bf1184927750df24ba8",
                "requestBody": {
                    "description": "Data for creating a new contact",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered contact data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/contact/{id}": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "List a contact by ID",
                "operationId": "783707de57724116780bdd03e441664d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Contact ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contact data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Contact"
                ],
                "summary": "Update a contact",
                "operationId": "b3b52c956445f19f048a7a31729d732f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Contact ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data for update contact",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated contact data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Contact"
                ],
                "summary": "Delete a contact",
                "operationId": "89f7327360c1e83ac0078616cdee160c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Contact ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ContactStatusEnum": {
                "description": "Contact status",
                "type": "string",
                "enum": [
                    "PENDING",
                    "WAITING_RESPONSE",
                    "ANSWERED",
                    "ARCHIVED"
                ]
            },
...
            "ApiErrorDTO": {
                "properties": {
                    "path": {
                        "type": "string",
                        "example": "/api/route"
                    },
                    "code": {
                        "type": "integer",
                        "example": 500
                    },
                    "message": {
                        "type": "string",
                        "example": "Error ocurried"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "email@gmail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "123"
                    }
                },
                "type": "object"
            },
...
            "ContactRequest": {
                "required": [
                    "name",
                    "email",
                    "subject",
                    "message",
                    "status"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 2,
                        "example": "João"
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 3,
                        "example": "joao@email.com"
                    },
                    "subject": {
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 3,
                        "example": "Oferta de emprego"
                    },
                    "message": {
                        "type": "string",
                        "minLength": 4,
                        "example": "<p>Boa tarde!</p><p>Tenho uma oferta de emprego para você</p>"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ContactStatusEnum"
                    }
                },
                "type": "object"
            },
...
            "Contact": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "João"
                    },
                    "email": {
                        "type": "string",
                        "example": "joao@email.com"
                    },
                    "subject": {
                        "type": "string",
                        "example": "Oferta de emprego"
                    },
                    "message": {
                        "type": "string",
                        "example": "<p>Boa tarde!</p><p>Tenho uma oferta de emprego para você</p>"
                    },
                    "status": {
                        "$ref": "#/components/schemas/ContactStatusEnum"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-28T11:23:31.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-28T11:23:31.000000Z"
                    }
                },
                "type": "object"
            },
...
            "AccessTokenDTO": {
                "properties": {
                    "access_token": {
                        "type": "string",
                        "example": "token123"
                    },
                    "token_type": {
                        "type": "string",
                        "example": "bearer"
                    },
                    "expires_in": {
                        "type": "number",
                        "example": 3600
                    }
                },
                "type": "object"
            }
        }
    },
...
}